requirements.txt
python-dotenv
sqlalchemy
asyncmy
wyoming
numpy
librosa
samplerate
aiosqlite
